pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wids-pi"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlmodel",
  "pydantic",
  "pyyaml",
  "loguru",
  "scapy",
]

[tool.setuptools.packages.find]
where = ["src"]
